requirements.txt
chromadb
langchain-text-splitters
requests
langchain_community
pypdf
flask
PyPDF2
langchain_core
python-dotenv
psycopg2-binary